+++ packages/xchain-bsc/__e2e__/bsc-client.e2e.ts
@@ -41,7 +41,7 @@ describe('xchain-evm (Bsc) Integration Tests', () => {
     expect(assetInfo).toEqual(correctAssetInfo)
   })
   it('should fetch bsc balances', async () => {
-    const address = clientTestnet.getAddress(0)
+    const address = await clientTestnet.getAddressAsync(0)
     console.log(address)
     const balances = await clientTestnet.getBalance(address)
     balances.forEach((bal: Balance) => {
@@ -88,14 +88,14 @@ describe('xchain-evm (Bsc) Integration Tests', () => {
   })
 
   it('should transfer 0.01 Bnb between wallet 0 and 1, with a memo', async () => {
-    const recipient = clientTestnet.getAddress(1)
+    const recipient = await clientTestnet.getAddressAsync(1)
     const amount = assetToBase(assetAmount('0.001', 18))
     const memo = `=:BNB.BNB:bnb1xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx:100000000000`
     const txHash = await clientTestnet.transfer({ amount, recipient, asset: AssetBNB, memo })
     console.log(txHash)
   })
   it('should transfer 0.001 eth between wallet 0 and 1, with a memo', async () => {
-    const recipient = clientTestnet.getAddress(1)
+    const recipient = await clientTestnet.getAddressAsync(1)
     const amount = assetToBase(assetAmount('0.001', 18))
 
     const memo = '=:BNB.BNB:bnb1xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx:100000000000'
@@ -138,7 +138,7 @@ describe('xchain-evm (Bsc) Integration Tests', () => {
   })
   it('should test estimates ', async () => {
     const estimateParams: EstimateApproveParams = {
-      fromAddress: clientTestnet.getAddress(0),
+      fromAddress: await clientTestnet.getAddressAsync(0),
       contractAddress: '0xd66c6b4f0be8ce5b39d52e0fd1344c389929b378', //ETh address
       spenderAddress: '0xdc4904b5f716Ff30d8495e35dC99c109bb5eCf81', //PancakeRouter contract on testnet
       amount: assetToBase(assetAmount('80', 18)),
@@ -147,7 +147,7 @@ describe('xchain-evm (Bsc) Integration Tests', () => {
     console.log(gasEstimate.toString())
     expect(gasEstimate.gte(0)).toBe(true)
 
-    const recipient = clientTestnet.getAddress(1)
+    const recipient = await clientTestnet.getAddressAsync(1)
     const amount = assetToBase(assetAmount('0.01', 18))
     const memo = '=:BNB.BUSD-BD1:bnb1xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx:100000000000'
     const gasEstimateWithMemo = await clientTestnet.estimateFeesWithGasPricesAndLimits({ amount, recipient, memo })