+++ packages/xchain-thorchain/src/client.ts
@@ -287,11 +287,7 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
   }
 
   /**
-   * Get the current address.
-   *
-   * @returns {Address} The current address.
-   *
-   * @throws {Error} Thrown if phrase has not been set before. A phrase is needed to create a wallet and to derive an address from it.
+   * @deprecated this function eventually will be removed use getAddressAsync instead
    */
   getAddress(index = 0): string {
     const address = this.cosmosClient.getAddressFromMnemonic(this.phrase, this.getFullDerivationPath(index))
@@ -300,6 +296,17 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
     }
 
     return address
+  }
+
+  /**
+   * Get the current address.
+   *
+   * @returns {Address} The current address.
+   *
+   * @throws {Error} Thrown if phrase has not been set before. A phrase is needed to create a wallet and to derive an address from it.
+   */
+  async getAddressAsync(walletIndex = 0): Promise<Address> {
+    return this.getAddress(walletIndex)
   }
 
   /**
@@ -348,7 +355,7 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
     const messageAction = undefined
     const offset = params?.offset || 0
     const limit = params?.limit || 10
-    const address = params?.address || this.getAddress()
+    const address = params?.address || (await this.getAddressAsync())
     const txMinHeight = undefined
     const txMaxHeight = undefined
 
@@ -446,20 +453,18 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
       return transaction
     } catch (error) {
       for (const fallback of FallBackUrls) {
-        for (const network of Object.keys(fallback)) {
-          try {
-            const networkObj = fallback[network as keyof typeof fallback]
-            const clientUrl = networkObj.node as string | string[]
-            const cosmosClient = new CosmosSDKClient({
-              server: Array.isArray(clientUrl) ? clientUrl[0] : clientUrl,
-              chainId: this.getChainId(network as Network),
-              prefix: getPrefix(network as Network),
-            })
-            const tx = await cosmosClient.txsHashGet(txId)
-            return tx
-          } catch (error) {
-            // Handle specific error if needed
-          }
+        try {
+          const networkObj = fallback[this.network]
+          const clientUrl = networkObj.node as string | string[]
+          const cosmosClient = new CosmosSDKClient({
+            server: Array.isArray(clientUrl) ? clientUrl[0] : clientUrl,
+            chainId: this.getChainId(this.network),
+            prefix: getPrefix(this.network),
+          })
+          const tx = await cosmosClient.txsHashGet(txId)
+          return tx
+        } catch (error) {
+          // Handle specific error if needed
         }
       }
       return null
@@ -692,7 +697,7 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
     const privKey = this.getPrivateKey(walletIndex)
     const signerPubkey = privKey.pubKey()
 
-    const fromAddress = this.getAddress(walletIndex)
+    const fromAddress = await this.getAddressAsync(walletIndex)
     const fromAddressAcc = cosmosclient.AccAddress.fromString(fromAddress)
 
     const depositTxBody = await buildDepositTx({
@@ -848,7 +853,7 @@ class Client extends BaseXChainClient implements ThorchainClient, XChainClient {
     }
 
     const txBody = await buildTransferTx({
-      fromAddress: this.getAddress(walletIndex),
+      fromAddress: await this.getAddressAsync(walletIndex),
       toAddress: recipient,
       memo,
       assetAmount: amount,