+++ packages/xchain-doge/CHANGELOG.md
@@ -1,8 +1,20 @@
-# v0.7.4 (2023-11-10)
+# v0.7.6 (2023-11-21)
 
 ## Update
 
 - BlOCKCYPHER_API_KEY renamed to BLOCKCYPHER_API_KEY
+
+# v0.7.5 (2023-11-16)
+
+## Update
+
+- Created method getAddressAsync
+
+# v0.7.4 (2023-11-10)
+
+## Update
+
+- Utxo-providers package from 0.2.5 to 0.2.6
 
 # v0.7.3 (2023-10-30)