+++ packages/xchain-midgard-query/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.1.6 (2023-11-14)
+
+## Update
+
+- update Midgard dep
+
 # v0.1.5 (2023-11-05)
 
 ## Update