+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.6.11 (2023-11-16)
+
+## Update
+
+- Created method getAddressAsync
+- update Midgard dep & thornode dep
+- Remove param fromAddress from quoteSwapParams to align with latest thornode specs 
+
 # v0.6.10 (2023-11-07)
 
 ## Fix